Warning
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@13510 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/int/arithmetic/nroot.hpp
@@ -191,7 +191,7 @@ namespace Gecode { namespace Int { namespace Arithmetic {
     GECODE_ME_CHECK(x0.gq(home,0));
     GECODE_ME_CHECK(x1.gq(home,0));
 
-    if (ops.exp() >= sizeof(int) * CHAR_BIT) {
+    if (static_cast<unsigned int>(ops.exp()) >= sizeof(int) * CHAR_BIT) {
       // The integer limits allow only 0 and 1 for x1
       GECODE_ME_CHECK(x1.lq(home,1));
     }
